Refactor code to use base test class
