--- tests/test_v1.py
@@ -1,515 +1,0 @@
-import unittest
-from base64 import b64encode
-
-from flask import json
-
-from app import create_app
-from app.init_db import db
-
-
-class TestMain(unittest.TestCase):
-    configurations = "testing"
-    app = create_app(configurations)
-
-    def setUp(self):
-        with self.app.app_context():
-            db.session.commit()
-            db.drop_all()
-            db.create_all()
-            self.client = self.app.test_client()
-            self.client.post("/v_1/register",
-                             data=json.dumps({"name": "tinyrick",
-                                              "password": "python8",
-                                              "email": "tinyrick@gmail.com"
-                                              }),
-                             content_type='application/json')
-            self.headers = {
-                'Authorization': 'Basic %s' %
-                                 b64encode(b"tinyrick@gmail.com:python8")
-                                     .decode("ascii")
-            }
-            res = self.client.get("/v_1/token", headers=self.headers)
-            token = json.loads(res.data)['token']
-            self.token_headers = {
-                'Authorization': 'Basic %s'
-                                 % b64encode(bytes(token + ':', "utf-8"))
-                                     .decode("ascii")
-            }
-
-    def test_user_registration(self):
-        """Test if a user is created using the api"""
-        response = self.client.post("/v_1/register",
-                                    data=json.dumps({"name": "esirick",
-                                                     "password": "morty11",
-                                                     "email": "esimorty@gmail.com"
-                                                     }),
-                                    content_type='application/json')
-        self.assertEqual(response.status_code, 201)
-        self.assertIn(
-            "Already Exists",
-            self.client.post("/v_1/register",
-                             data=json.dumps({"name": "esirick",
-                                              "password": "python8",
-                                              "email": "esimorty@gmail.com"
-                                              }),
-                             content_type='application/json').data.decode())
-        response_no_arg = self.client.post("/v_1/register",
-                                           data=json.dumps({"name": "esirick",
-                                                            "password": "python8",
-                                                            "email": "  "
-                                                            }),
-                                           content_type='application/json')
-        """
-        Test if no arguments are passed the service aborts
-        """
-        self.assertEqual(400, response_no_arg.status_code)
-
-    def test_verify_password(self):
-        """
-        Test if user is authenticated using email and password
-        """
-        res = self.client.get("/v_1/token", headers=self.headers)
-        self.assertEqual(res.status_code, 200)
-
-    def test_authentication_with_token(self):
-        """
-        Test if user is authenticated using a valid token
-        """
-        res = self.client.get("/v_1/token", headers=self.token_headers)
-        """Accessing Secure endpoint using a token"""
-        self.assertEqual(res.status_code, 200)
-
-    def test_user_login(self):
-        """
-        Test If A Registered User Can Login to the app using registration
-         credentials
-        """
-        self.client.post(
-            "/v_1/register",
-            data=json.dumps({"name": "esirick",
-                             "password": "morty00",
-                             "email": "mortymorty@gmail.com"
-                             }),
-            content_type='application/json')
-        response = self.client.post(
-            "/v_1/user",
-            data=json.dumps({
-                "password": "morty00",
-                "email": "mortymorty@gmail.com"
-            }),
-            content_type='application/json')
-        self.assertEqual(200, response.status_code)
-
-    def test_user_can_not_login_with_invalid_creds(self):
-        """
-        Test If A Registered User Can Login to the app using
-        Invalid Credentials
-         """
-        # email "tinyrick@gmail.com" was created on setup but with 'python8' as the pass
-        response = self.client.post(
-            "/v_1/user",
-            data=json.dumps({
-                "password": "morty00",
-                "email": "tinyrick@gmail.com"
-            }),
-            content_type='application/json')
-        self.assertEqual(401, response.status_code)
-
-    def test_un_registered_user_can_not_login(self):
-        """
-        Test If a Un Registered User Can Login to the app
-         """
-        response = self.client.post(
-            "/v_1/user",
-            data=json.dumps({
-                "password": "morty00",
-                "email": "notregistered@gmail.com"
-            }),
-            content_type='application/json')
-        self.assertEqual(401, response.status_code)
-
-    def test_add_shopping_list(self):
-        """
-        Test a Logged in User can add a shopping list into his/her account
-        """
-        response = self.client.post(
-            "/v_1/shoppinglists",
-            data=json.dumps({
-                "name": "Soko",
-                "description": "Short Description About Soko"
-            }),
-            content_type='application/json', headers=self.headers)
-        self.assertEqual(201, response.status_code)
-
-    def test_view_shopping_lists(self):
-        """
-        Test a logged in user can view all his/her shoppinglists
-         together with the items in them 
-        """
-        self.create_shopping_lists("Family")
-        self.create_shopping_lists("Games")
-
-        self.create_shopping_lists_item("Wine", 2500, 3, "Family")
-        self.create_shopping_lists_item("Ball", 2500, 3, "Games")
-
-        response = self.client.get(
-            "/v_1/shoppinglists",
-            headers=self.headers)
-        self.assertEqual(2, len(json.loads(response.data)))
-
-        x = json.loads(response.data)
-
-        self.assertEqual(1, len(x[0]['items']))
-
-    def test_user_can_query_for_a_shopping_list(self):
-        """
-                Test a logged in user can search for a shoppinglists
-                """
-        self.create_shopping_lists("Family")
-        self.create_shopping_lists_item("Wine", 2500, 3, "Family")
-        response = self.client.get(
-            "/v_1/shoppinglists?q=Family",
-            headers=self.headers)
-        self.assertEqual(200, response.status_code)
-
-    def test_shopping_list_name_is_unique(self):
-        """
-        Test shopping_list cannot be added more than once
-        """
-        self.create_shopping_lists("Sons Birthday")
-
-        response = self.client.post(
-            "/v_1/shoppinglists",
-            data=json.dumps({
-                "name": "Sons Birthday",
-                "description": "Short Description About Sons Birthday"
-            }),
-            content_type='application/json', headers=self.headers)
-        self.assertEqual(409, response.status_code)
-
-    def test_unauthenticated_user_cannot_create_shopping_list(self):
-        """
-        Test a non Logged in User can not add a shopping list
-        """
-        response = self.client.post(
-            "/v_1/shoppinglists",
-            data=json.dumps({
-                "name": "Soko",
-                "description": "Short Description About Soko"
-            }),
-            content_type='application/json')
-        # 401 means user is un authenticated
-        self.assertEqual(401, response.status_code)
-
-    def test_update_shopping_list(self):
-        """
-        Test a Logged in User can Update his/her shopping lists
-        """
-        self.create_shopping_lists("Movies")
-        response = self.client.put(
-            "/v_1/shoppinglists",
-            data=json.dumps({
-                "name": "Movies",
-                "new_name": "Series",
-                "description": "None"
-            }),
-            content_type='application/json', headers=self.headers)
-        self.assertEqual(200, response.status_code)
-
-    def test_update_of_non_existing_shopping_list_fails(self):
-        """
-        Test a Logged in User can not Update a non existing shopping lists
-        """
-        response = self.client.put(
-            "/v_1/shoppinglists",
-            data=json.dumps({
-                "name": "I Do not Exist",
-                "new_name": "I Exist",
-                "description": "None"
-            }),
-            content_type='application/json', headers=self.headers)
-        self.assertEqual(404, response.status_code)
-
-    def test_delete_of_a_shopping_list(self):
-        """
-        Test a Logged in User Can Delete His/Her shopping lists 
-        """
-        self.create_shopping_lists("Gaming")
-        response = self.client.delete(
-            "/v_1/shoppinglists",
-            data=json.dumps({
-                "name": "Gaming"
-            }),
-            content_type='application/json', headers=self.headers)
-        self.assertEqual(200, response.status_code)
-
-    def test_delete_of_non_existing_shopping_list_fails(self):
-        """
-        Test a Logged in User Can not Delete a non existing shopping lists 
-        """
-        response = self.client.delete(
-            "/v_1/shoppinglists",
-            data=json.dumps({
-                "name": "Not existing"
-            }),
-            content_type='application/json', headers=self.headers)
-        self.assertEqual(404, response.status_code)
-
-    def test_non_user_can_not_delete_a_shopping_list(self):
-        """
-        Test a Non Logged in User Can not Delete a shopping lists 
-        """
-        self.create_shopping_lists("Morties")
-        response = self.client.delete(
-            "/v_1/shoppinglists",
-            data=json.dumps({
-                "name": "Morties"
-            }),
-            content_type='application/json')  # No headers provided
-        self.assertEqual(401, response.status_code)
-
-    def test_user_can_share_shopping_list(self):
-        """
-        Test a Logged in User Can Share a shopping list with other users
-        """
-        self.create_shopping_lists("Shares")
-        self.client.post("/v_1/register",
-                         data=json.dumps({"name": "tinyrick",
-                                          "password": "python8",
-                                          "email": "esir@gmail.com"
-                                          }),
-                         content_type='application/json')
-
-        response = self.client.post(
-            "/v_1/shoppinglists/share",
-            data=json.dumps({
-                "name": "Shares",
-                "email": "esir@gmail.com"
-            }),
-            content_type='application/json', headers=self.headers)
-        self.assertEqual(200, response.status_code)
-        headers2 = {
-            'Authorization': 'Basic %s' %
-                             b64encode(b"esir@gmail.com:python8")
-                                 .decode("ascii")
-        }
-        res = self.client.get(
-            "/v_1/shoppinglists",
-            data=json.dumps({
-                "name": "Shares",
-                "email": "esir@gmail.com"
-            }),
-            content_type='application/json', headers=headers2)
-        self.assertEqual(1, len(json.loads(res.data)))
-
-    def test_shopping_list_cannot_be_shared_with_non_app_users(self):
-        self.create_shopping_lists("Shares")
-        response = self.client.post(
-            "/v_1/shoppinglists/share",
-            data=json.dumps({
-                "name": "Shares",
-                "email": "amnotthere@gmail.com"
-            }),
-            content_type='application/json', headers=self.headers)
-        self.assertEqual(404, response.status_code)
-
-    def test_non_existing_shopping_list_cannot_be_shared(self):
-        response = self.client.post(
-            "/v_1/shoppinglists/share",
-            data=json.dumps({
-                "name": "Shares",
-                "email": "esir@gmail.com"
-            }),
-            content_type='application/json', headers=self.headers)
-        self.assertEqual(404, response.status_code)
-
-    def test_shopping_list_item_created_successfully(self):
-        """
-        Test a Logged In User Can Add items to their shopping_lists
-        """
-        self.create_shopping_lists("School")
-        response = self.create_shopping_lists_item("Skuma", 10, 2, "School")
-        self.assertEqual(201, response.status_code)
-
-    def test_existing_shopping_list_item_can_not_be_created(self):
-        """
-        Test a Logged In User Can not Add an item more than once to 
-        their shopping_lists
-        """
-        self.create_shopping_lists("Holiday")
-        self.create_shopping_lists_item("Router", 2500, 24, "Holiday")
-        response = self.create_shopping_lists_item("Router", 2500, 24, "Holiday")
-        self.assertEqual(409, response.status_code)
-
-    def test_shopping_list_not_found_returned(self):
-        """
-        Test 404 is returned on attempt to add items on non existing shoppinglist 
-        :return: 
-        """
-        response = self.create_shopping_lists_item("Wounder Woman", 150, 2, "Movies")
-        self.assertEqual(404, response.status_code)
-
-    def test_un_authenticated_users_cannot_add_items(self):
-        """
-        Test non registered users not allowed to add shoppinglist items 
-        """
-        response = self.client.post(
-            "/v_1/shoppinglist_items",
-            data=json.dumps({
-                "name": "Amazing Woman",
-                "price": "150",
-                "quantity": "2",
-                "shopping_list_name": "Movies"
-            }
-            ),
-            content_type='application/json')  # authentication headers not passed
-        self.assertEqual(401, response.status_code)
-
-    def test_updating_shopping_list_items(self):
-        """
-        Test a Logged in User can update their shopping list items successfully
-        """
-        self.create_shopping_lists("Party")
-        self.create_shopping_lists_item("Beer", 250, 24, "Party")
-
-        response = self.client.put(
-            "/v_1/shoppinglist_items",
-            data=json.dumps({
-                "name": "Beer",
-                "new_name": "BeerX",
-                "price": 0,
-                "quantity": 0,
-                "shopping_list_name": "Party",
-                "new_shopping_list_name": "None"
-            }
-            ),
-            content_type='application/json', headers=self.headers)
-        self.assertEqual(200, response.status_code)
-
-    def test_user_can_not_update_non_existing_item(self):
-        """
-        Tests if a User Can Update a Non-existing item in their shopping lists
-        """
-        self.create_shopping_lists("Fictions")
-        response = self.client.put(
-            "/v_1/shoppinglist_items",
-            data=json.dumps({
-                "name": "Flash",
-                "new_name": "Flask",
-                "price": 0,
-                "quantity": 0,
-                "shopping_list_name": "Fiction",
-                "new_shopping_list_name": "None"
-            }
-            ),
-            content_type='application/json', headers=self.headers)
-        self.assertEqual(404, response.status_code)
-
-    def test_user_can_not_update_item_to_non_existing_shopping_list(self):
-        """
-        Tests updating an item in one shopping list to a shopping list
-        that does not exist fails
-        """
-        self.create_shopping_lists("Back to school")
-        self.create_shopping_lists_item("Beer", 250, 24, "Back to school")
-
-        response = self.client.put(
-            "/v_1/shoppinglist_items",
-            data=json.dumps({
-                "name": "Beer",
-                "new_name": "Alcohol",
-                "price": 0,
-                "quantity": 0,
-                "shopping_list_name": "Back to school",
-                "new_shopping_list_name": "Fiction"
-            }
-            ),
-            content_type='application/json', headers=self.headers)
-        self.assertEqual(404, response.status_code)
-
-    def test_delete_shopping_list_item(self):
-        """
-        tests if a logged in user is able to delete a shopping list item
-        from his/her shopping lists
-        """
-        self.create_shopping_lists("BBQ")
-        self.create_shopping_lists_item("Meat", 4000, 20, "BBQ")
-        response = self.delete_shopping_list_item(name="Meat", shopping_list_name="BBQ")
-        self.assertEqual(200, response.status_code)
-
-    def test_delete_of_non_existing_shopping_list_item_fails(self):
-        """
-        tests response if a user tries to delete non-existing item
-        """
-        self.create_shopping_lists("Love")
-        response = self.delete_shopping_list_item(name="Meat", shopping_list_name="Love")
-        self.assertEqual(404, response.status_code)
-
-    def test_user_cannot_delete_an_item_from_non_existing_shopping_list(self):
-        """
-        Tests that a user is not able to delete an item from a shopping list
-        that has not been created
-        :return: 
-        """
-        response = self.delete_shopping_list_item("Siko", "Pia Siko")
-        self.assertEqual(404, response.status_code)
-
-    def test_non_authenticated_user_cannot_delete_an_item(self):
-        response = self.client.delete(
-            "/v_1/shoppinglist_items",
-            data=json.dumps({
-                "name": "any name",
-                "shopping_list_name": "Try me"
-            }),
-            content_type='application/json')  # No Headers
-        self.assertEqual(401, response.status_code)
-
-    def create_shopping_lists(self, name):
-        """
-        Creates Shopping list with the Provided name
-        :param name: 
-        """
-        self.client.post(
-            "/v_1/shoppinglists",
-            data=json.dumps({
-                "name": "" + name,
-                "description": "Short Description About " + name + "Shopping List"
-            }),
-            content_type='application/json', headers=self.headers)
-
-    def create_shopping_lists_item(self, name, price, quantity, shopping_list_name):
-        """
-        Creates a shopping list item with provided details
-        :param name: Of the item
-        :param price: of the item
-        :param quantity: of the item
-        :param shopping_list_name: to which the item belongs to
-        :return: a response
-        """
-        return self.client.post(
-            "/v_1/shoppinglist_items",
-            data=json.dumps({
-                "name": name,
-                "price": price,
-                "quantity": quantity,
-                "shopping_list_name": shopping_list_name,
-            }
-            ),
-            content_type='application/json', headers=self.headers)
-
-    def delete_shopping_list_item(self, name, shopping_list_name):
-        """
-        Deletes a given shopping list item
-        :return: 
-        """
-        return self.client.delete(
-            "/v_1/shoppinglist_items",
-            data=json.dumps({
-                "name": name,
-                "shopping_list_name": shopping_list_name
-            }),
-            content_type='application/json', headers=self.headers)
-
-
-if __name__ == '__main__':
-    unittest.main()
